tests/submission/Microsoft/semantic/aside.html

<!DOCTYPE html>
<html><head>
 <meta content="text/html; charset=utf-8" http-equiv="content-type">
 <title>verify ASIDE is a HTMLElement</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
<link href="http://www.w3.org/TR/html5/the-aside-element.html#the-aside-element" rel="help">
</head>
<body>
<div id="log"></div>
 <aside id="aside">
 This is an ASIDE with a NAV in it
 <nav>
 <ul>
  <li><a href="#">some test link</a>
  </li><li><a href="#">another test link</a>
 </li></ul>
 </nav>
</aside>
<script type="text/javascript">
 test(function()
 {
    assert_array_equals( "[object HTMLElement]",
                         document.getElementById("aside").toString() )
 })
</script>


</body></html>
